smokeview source: remove pp_FRAME directive, move frame routines to IOframe.c/IOframe.h
