+++ Source/smokeview/options.h
@@ -25,7 +25,6 @@
 #define pp_GPUTHROTTLE      // pp_GPU directive must also be set
 #endif
 #define pp_NOBOUNDS           // add option to not compute global bounds at startup
-//#define pp_FRAME          // generalized routines to read in frame data
 
 //*** options: windows
