+++ Source/smokeview/readsmv.c
@@ -5853,9 +5853,6 @@ int ParseSLCFProcess(int option, bufferstreamdata *stream, char *buffer, int *nn
   sd->loadstatus = FILE_UNLOADED;
 #endif
   sd->slice_mask       = NULL;
-#ifdef pp_FRAME
-  sd->frameinfo        = NULL;
-#endif
   sd->vals2d.vals      = NULL;
   sd->vals2d.vals_orig = NULL;
   sd->vals2d.times     = NULL;
@@ -6027,9 +6024,6 @@ int ParseSLCFProcess(int option, bufferstreamdata *stream, char *buffer, int *nn
     sd->slicelabel = SLICEBUFFER(lenslicelabel);
     strcpy(sd->slicelabel, slicelabel);
   }
-#ifdef pp_FRAME
-  sd->frameinfo = FRAMEInit(sd->file, sd->size_file, GetSliceFrameInfo);
-#endif
   if(read_slice_header==1){
     int error;
