+++ Source/smokeview/structures.h
@@ -1473,9 +1473,6 @@ typedef struct _slicedata {
   FILE_SIZE file_size;
   int *geom_offsets;
   devicedata vals2d;
-#ifdef pp_FRAME
-  framedata *frameinfo;
-#endif
 #ifdef pp_SLICE_MULTI
   int loadstatus;
 #endif
